Moving internal dependencies into an Internal namespace
Trying this out to keep it clean as the project author. Shouldn't make a difference to any users.
