+++ src/GitModel/CommitFileReader.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using GitModel.Internal;
 
 namespace GitModel
 {
